BUG-44248: Fix iotas jar shading issues and add logging.
When shading for com.fasterxml.jackson.databind was added to storm
module, the parser class could not be instantiated. Reason was that
ObjectMapper it was using was not packaged in parser jar. Hence had to
add maven-shade-plugin to parser/pom.xml. Once that was added, it was
clashing with storm cores jackson classes. Hence shading in
parsers/pom.xml. Packaging a fat jar for parsers module now was also
bringing in hadoop-client and slf4j-api and slf4j-log4j12. These
dependencies were relying on some other jars to be available at runtime
that was causing issues. Hence they either had to excluded or removed
from dependency.