STREAMLINE-434: Fix bug introduced due to verifying CP impl.
As part of STREAMLINE-434 I added check to verify CP upload such that
the jar has the implementation class specified in CustomProcessorInfo.
However, that made the input stream of the jar reach the end. Hence
when it was copied to local file system storage it created a file of
size 0.