Github #483 - Docker Support
Initial support for building Docker images via Maven. The image can be
built as part of the build process by using the Maven Profile
“-Pdocker” and does expect that “streamline-dist” has been previously
built. The easiest way to ensure this is always including “-Pdist” with
the Maven command as well. I'm not sure of this teams contribution
guidelines so I apologize if you want a Github issue or JIRA attached
to this let me know and I'll be glad to add one. This was the same
approach we used for NiFi and find it pretty handy. This PR is for
Github Issue #483