ISSUE-765: DB based file storage
Currently, we expect users to use HDFS as the FileStorage incase of HA, so that
users uploading jar from one server connection can be seen on another server.

Storing the jars as a blob into DB as a workaround lets us do away with HDFS dependency for HA. This will delegate
the replication to the db and if users need HA for stored files, they can have a DB cluster
with some kind of a master-master or a master-slave replication.