ISSUE-829: Fix for Concat function
CONCAT was assumed to be a calcite built-in function, however the calcite parser
returns it as a unresolved user defined function.
Due to this, to topology deployment code tries to download and package the jar which fails.

Also commenting the SUBSTRING function for now since theres no way from UI to pass the start index field as parameter for the Function.