ISSUE-933: SAM Processors/Sinks/Rules need to update when its upstream components are updated
Whenever an edge or an output stream is updated or deleted, the downstream processor/sink/rule to which it is connected will be marked for 'reconfiguration' by the backend. UI can use this field to prompt the user to reconfigure the component. Once the user takes some action and saves the updated component, the reconfigure flag is reset by the backend.

During deployment if any component or rule has 'reconfigure' set, the deployment will fail.

Author: Arun Mahadevan <arunm@apache.org>

Closes #934 from arunmahadevan/ISSUE-933