ISSUE-968 Join processor output when both streams have same field
Join processor support for multiple streams with same fields name by using a unique alias.
the screenshot below:

![joinscreenshot](https://user-images.githubusercontent.com/6010694/32105010-8a431c32-bb44-11e7-8da7-ebe034bbb834.png)

Stream legends are used to denote which field belong to which stream.
We can have in the way it’s shown in the screenshot or add an icon beside the “Output Fields*” label where a user can hover and check.
I am preferring the former approach so the user doesn’t have to hover over the icon every now and then to verify as it will always be available on the screen.

Open for a suggestion from others.

shahsank3t  arunmahadevan  roshannaik please review and comment.
Thanks in advanced

Author: sameer79 <findsameershaikh@yahoo.co.in>

Closes #991 from sameer79/ISSUE-968

Change-Id: I75395dbaa3700bc1c4c2a9ba7f4544790ac280a3