ISSUE-1260: Changes to discover atlas and enable storm hook to notify
Discover and enable Storm Atlas notifier plugin if user has selected 'Atlas' in the environment.

This patch does not handle shipping the necessary atlas jars in the storm path. This can be handled in a separate patch by optionally shipping relevant atlas jars via streamline or have Ambari/Atlas to install the storm atlas hook jars in the HDF cluster.

Closes #1260

Author: Arun Mahadevan <arunm@apache.org>

Closes #1261 from arunmahadevan/ISSUE-1260