ISSUE-1313 Not able to add rule in Rule processor if we use alphanumeric column name
sameer79 - Was there any particular need for adding 0-9 into this regex? Seems like that's causing the issue. Could you please clarify if it's good to remove the number or is needed? Thanks.

arunmahadevan - I have tested with the basic query and works fine for me now. Could you please try some complex query that can be run by the user so we can be sure of this fix? Thanks.

Author: Sanket Shah <er.sanketshah@gmail.com>

Closes #1315 from shahsank3t/ISSUE-1313