ISSUE-1338 Exclude Avro dependency from all possible client paths to not conflict with SR's Avro version
SR depends on Avro 1.8 whereas dependencies in Hadoop ecosystem depend on Avro 1.7 hence conflicted.
This patch fixes the issue to try best to leverage SR's Avro version in runtime topology.

This depends on https://github.com/hortonworks/registry/pull/541 - so we need to update SR's version here in the patch. I just didn't do this because it requires new release of SR.

Author: Jungtaek Lim (HeartSaVioR) <kabhwan@gmail.com>

Closes #1339 from HeartSaVioR/ISSUE-1338